🔥 chore: move lib to src
