+++ src/components/ButtonLogin.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 import GitHubIcon from "@mui/icons-material/GitHub";
 
 import { Button } from "./@shared";
-import { GITHUB_LOGIN_OAUTH_URI } from "lib/config";
+import { GITHUB_LOGIN_OAUTH_URI } from "src/config";
 
 function ButtonLogin() {
   return (
