+++ src/components/modal/ModalDeleteConfirm.tsx
@@ -3,10 +3,10 @@ import { setCookie } from "cookies-next";
 import { Box, Typography } from "@mui/material";
 
 import { Button } from "../@shared";
-import { deleteUserDeployment } from "lib/api";
-import useModal from "lib/hooks/useModal";
-import loginState from "lib/recoil/auth";
-import { userDeploymentsState } from "lib/recoil/userDeployments";
+import { deleteUserDeployment } from "src/api";
+import useModal from "src/hooks/useModal";
+import loginState from "src/recoil/auth";
+import { userDeploymentsState } from "src/recoil/userDeployments";
 
 import { UserDeploymentData } from "types/deployment";
 import { LoginData } from "types/auth";
